Add files via upload
This project showcases a web development module assignment for our university. It includes a well-designed login form and registration form implemented using HTML and CSS. The forms are intended to provide a user-friendly and visually appealing experience for users accessing the website. The repository serves as a demonstration of front-end development skills and serves as a foundation for potential further enhancements and features.